Added Initial level Cards functionality
Created these classes:
1. Card
2. BallDeliveryEvent

Card class is assigned to player. each card will contain a minimum of
one BallDeliveryEvent class to a maximum of 6
The BallDeliveryEvent class basically has a small program within it
that will be executed on that ball. details on this class is commented
within the class.

PENDING:
using the cards in the actual calculation flow.